add support for fp8 kv cache using flash infer

+++ server/lorax_server/models/flash_causal_lm.py
@@ -963,12 +963,12 @@ class FlashCausalLM(Model):
                 (
                     torch.empty(
                         (num_blocks, BLOCK_SIZE, num_heads, head_size),
-                        dtype=dtype,
+                        dtype=torch.float8_e4m3fn,
                         device=device,
                     ),
                     torch.empty(
                         (num_blocks, BLOCK_SIZE, num_heads, head_size),
-                        dtype=dtype,
+                        dtype=torch.float8_e4m3fn,
                         device=device,
                     ),
                 )
